# core: Isolate should abstract shared buffer