--- a/BUILD.gn
+++ b/BUILD.gn
@@ -13,8 +13,7 @@ group("default") {
     ":deno",
     ":hyper_hello",
     ":test_rs",
-    "core:deno_core_http_bench",
-    "core:deno_core_test",
+    "core:default",
     "libdeno:test_cc",
   ]
 }