We always pass around Box<[u8]>, and adding this generic is an
unnecessary complication.

<!--
Before submitting a PR read https://deno.land/manual.html#contributing
-->