@ry 2019-03-15
This branch has about a 20% slow down in deno_core_http_bench

master:
```
~/src/deno ./third_party/wrk/mac/wrk -c 10 -d 10 http://127.0.0.1:4544/
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   700.79us    2.72ms  50.11ms   95.29%
    Req/Sec    27.68k     2.31k   32.47k    74.26%
  556208 requests in 10.10s, 27.05MB read
Requests/sec:  55069.14
Transfer/sec:      2.68MB
```
this branch:
```
~/src/deno ./third_party/wrk/mac/wrk -c 10 -d 10 http://127.0.0.1:4544/
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.89ms    8.82ms 176.84ms   95.67%
    Req/Sec    22.34k     2.50k   30.35k    72.28%
  448968 requests in 10.10s, 21.84MB read
Requests/sec:  44454.64
Transfer/sec:      2.16MB
```