@ry 2019-03-15
slowdown reduced to 9%

master branch:
```
~/src/deno ./third_party/wrk/mac/wrk -c 10 -d 10 http://127.0.0.1:4544/
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   496.83us    1.81ms  33.27ms   95.89%
    Req/Sec    26.97k     2.28k   35.55k    74.63%
  539459 requests in 10.10s, 26.24MB read
Requests/sec:  53414.27
Transfer/sec:      2.60MB
```
this branch:
```
~/src/deno ./third_party/wrk/mac/wrk -c 10 -d 10 http://127.0.0.1:4544/
Running 10s test @ http://127.0.0.1:4544/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   416.97us    1.29ms  26.02ms   95.85%
    Req/Sec    24.18k     2.40k   30.03k    71.78%
  486153 requests in 10.10s, 23.65MB read
Requests/sec:  48135.23
Transfer/sec:      2.34MB
```