@piscisaureus 2019-03-15
> slowdown reduced to 9%

So this shared queue is slower than the current one ?