@ry 2019-03-15
> So this shared queue is slower than the current one ?

It's 9% slower than master branch's core/http_bench.js 
https://github.com/denoland/deno/blob/45fad1b7cfd44376d8be35288b7213636bc93450/core/http_bench.js
It's doing slightly more work - it appears very sensitive in this range - I guess it's just that we're using slightly more JS - with a few more ArrayBuffer constructors. I think we have to take the hit here in order to support this more general API. 
Hopefully we can make up for it elsewhere.