@piscisaureus 2019-03-15
Rather than converting to a Box<[u8]>, wouldn't it be possible to return a borrowed subslice?
I think the slowdown could potentially originate from excessive memory allocations.