@ry 2019-03-15
@piscisaureus libmalloc is showing up at about 20% in the profile - but i'm not sure how that compares to normal v8 loads
<img width="1096" alt="Screen Shot 2019-03-15 at 3 30 36 AM" src="https://user-images.githubusercontent.com/80/54415464-bbd6be00-46d2-11e9-9a4b-6074c4d0f614.png">

these are tiny 16 byte messages... i find it hard to imagine they're causing that much perf drop