@piscisaureus 2019-03-15
> It's 9% slower than master branch's core/http_bench.js

I don't understand how this can be slower whereas core_improvements2 is twice as fast as master.
What are we missing here?