@ry 2019-03-15
> I don't understand how this can be slower whereas core_improvements2 is twice as fast as master.
What are we missing here?

We've landed all the //core changes from core_improvements2 - so as far as //core/http_bench.js - it's the same code pretty much.