@piscisaureus 2019-03-15
> [02:54, 3/15/2019] piscisaureus: I'm curious how the two branches  compare
> [02:54, 3/15/2019] piscisaureus: Did we make any progress at all
> [02:54, 3/15/2019] Ryan Dahl: vs master?
> [02:54, 3/15/2019] piscisaureus: Yeah
> [02:55, 3/15/2019] Ryan Dahl: using v0.3.3 .. yes it's about 2x better
> [02:55, 3/15/2019] piscisaureus: Oh ok!
> [02:55, 3/15/2019] piscisaureus: That's encouraging
> [02:56, 3/15/2019] Ryan Dahl: 17k vs 31k

Maybe I was misreading this, and you meant v0.3.3 already includes the 2x performance increase.