@ry 2019-03-15
> Maybe I was misreading this, and you meant v0.3.3 already includes the 2x performance increase.

That was regarding "deno tests/http_bench.ts" ... this is about "deno_core_http_bench" (aka //core/http_bench.js)

The naming is super confusing. We need to rename soon.