@ry 2019-03-15
Getting this failure on arm64
```
$CARGO_TARGET_DIR/aarch64-unknown-linux-gnu/release/deno tests/002_hello.ts
qemu: Unsupported syscall: 278
qemu: Unsupported syscall: 278
```
cc @afinch7 (any idea?)