@ry 2019-03-15
@piscisaureus I'm getting this CI failure on windows - any idea?
```
FAILED: deno_core_http_bench_test.exe deno_core_http_bench_test.exe.pdb 
ninja -t msvc -e environment.x64 -- ../../third_party/llvm-build/Release+Asserts/bin/lld-link.exe /nologo /OUT:./deno_core_http_bench_test.exe /PDB:./deno_core_http_bench_test.exe.pdb @./deno_core_http_bench_test.exe.rsp
lld-link: error: <root>: undefined symbol: mainCRTStartup
lld-link: error: undefined symbol: __CxxFrameHandler3
>>> referenced by rust_crates/deno_core_http_bench_test_bin.o:(.xdata)
lld-link: error: undefined symbol: __CxxFrameHandler3
>>> referenced by libtest-70b4d6fe9ba4c17b.rlib(test-70b4d6fe9ba4c17b.test.7h7r2drh-cgu.0.rcgu.o):(.xdata)
>>> referenced by libtest-70b4d6fe9ba4c17b.rlib(test-70b4d6fe9ba4c17b.test.7h7r2drh-cgu.0.rcgu.o):(.xdata)
>>> referenced by libtest-70b4d6fe9ba4c17b.rlib(test-70b4d6fe9ba4c17b.test.7h7r2drh-cgu.0.rcgu.o):(.xdata)
>>> referenced by libtest-70b4d6fe9ba4c17b.rlib(test-70b4d6fe9ba4c17b.test.7h7r2drh-cgu.0.rcgu.o):(.xdata)
>>> referenced by libtest-70b4d6fe9ba4c17b.rlib(test-70b4d6fe9ba4c17b.test.7h7r2drh-cgu.0.rcgu.o):(.xdata)
>>> referenced by libtest-70b4d6fe9ba4c17b.rlib(test-70b4d6fe9ba4c17b.test.7h7r2drh-cgu.0.rcgu.o):(.xdata)
```