@piscisaureus 2019-03-15
> @piscisaureus I'm getting this CI failure on windows - any idea?

hmm