@piscisaureus 2019-03-15
Obviously it is missing msvcrt symbols. But why isn't it linking the crt... :/