@ry 2019-03-15
Here's what the 9% slowdown looks like on the benchmarks page now that it's landed:
<img width="699" alt="Screen Shot 2019-03-15 at 11 41 05 AM" src="https://user-images.githubusercontent.com/80/54443666-712b6500-4717-11e9-99d2-7cd829168414.png">